refactor: GlobalState owns what only needs GlobalState; one prepare step for firing a pointer; call-outs addressed by ID
`takeover`, `prioritize_call_out`, `committer_stats` and the `initialize_*`
family move from `Vm` to `GlobalState` (`self: &Arc<Self>` where a template is
minted). `Vm` keeps `pub` forwarders for the initializers the tests, benches
and `lpcc` call, now on `&self`. `send_op`, the private `tx`/`config`
forwarders, `new_task_template` (`TaskTemplate::from(global_state)` field for
field) and two commented-out remnants are deleted.

`GlobalState::prepare_function_ptr` is the resolve, `triple` and
init-if-needed sequence that `prioritize_call_out` and telnet's
`resolve_input_to` each carried; init failure is reported inside it and
surfaces as `Ok(None)`. Each site keeps its own run step and error reporting.

The call-out timer now sends the call out's ID; the queue's slot family
(`get`/`get_mut`/`remove`, `next_push_index`) is gone and `push` is a test
fixture. `cancelled_call_outs` is a `HashSet`.

+++ src/interpreter/vm/vm_op.rs
@@ -13,8 +13,8 @@ pub enum VmOp {
     /// Start the login process for a connection.
     InitiateLogin(Arc<Connection>),
 
-    /// Run a CallOut function, identified by its index in the [`CallOuts`](crate::interpreter::call_outs::CallOuts) list
-    PrioritizeCallOut(usize),
+    /// Run a scheduled [`CallOut`](crate::interpreter::call_outs::CallOut), identified by its ID.
+    PrioritizeCallOut(u64),
 
     /// The Task with the passed ID has failed.
     TaskError(TaskId, Box<LpcError>),
